[!] Improve Authenticator implementation
The Authenticator now supports and interface, a base class, and internally handles the deletion of AuthTokens.

Added:
- Interface `AuthenticatorInterface` and class `AbstractAuthenticator`
- Method `AbstractAuthenticator::isLoggedOut()` to indicate if the logout method has been called
- Method `AbstractAuthenticator::getAuthenticationToken()` to return the `AuthToken` if user is authenticated by tokens
- Method `AbstractAuthenticator::clearAuthenticator()` to clear internal values
- Methods `userId()` and `getUserId()` to quickly retrieve the user ID of the current user
- Methods on `AbstractAuthenticator` to delete user data from session, cookies, and auth tokens

Changed:
- Class `AbstractAuthenticator` to implement `AuthenticatorInterface`
- Renamed class constants `METHOD_*` with `AUTH_BY_*`
- Renamed method `isAuthenticated()` to `check()` on `AbstractAuthenticator`
- Renamed method `updateCookie()` to `updateCurrentToken()` on `AbstractAuthenticator`
- Methods `login()` and `logout()` to return void
- Method `logout()` to use new `deleteUserSession()` and `deleteUserTokens()` methods

Removed:
- Removed class property `$isAuthenticated` from `AbstractAuthenticator` (replaced with `$authenticatedUser` property)

Notes:
- `AbstractAuthenticator` is designed for `AuthenticatableInterface` + `ModelInterface`
- `Authenticator` is designed for `UserInterface` and `ContentInterface`
